- Parts now have only the number of pitches needed - Time signatures as one hot vectors - All pieces are transposed to C/Am during training and inference   - No data augmentation needed anymore

+++ data.py
@@ -21,27 +21,62 @@ indices_parts = {
 
 indices_extra = {
     'has_fermata': 0,
-    'num_sharps': 1,
-    'time_numerator': 2,
-    'time_denominator': 3,
+    'has_time_signature_3/4': 1,
+    'has_time_signature_4/4': 2,
+    'has_time_signature_3/2': 3,
     'time_pos': 4
 }
 
-pitch_size = 60
-part_size = pitch_size + len(indices_parts)
+sharps_to_offset = {
+    -4: -4,
+    -3: 3,
+    -2: -2,
+    -1: 5,
+    0: 0,
+    1: -5,
+    2: 2,
+    3: -3,
+    4: 4
+}
+
+offset_to_sharps = {
+    -5: 1,
+    -4: -4,
+    -3: 3,
+    -2: -2,
+    0: 0,
+    2: 2,
+    3: -3,
+    4: 4,
+    5: -1
+}
+
+offsets_parts = {
+    'soprano': 55,
+    'alto': 49,
+    'tenor': 45,
+    'bass': 31
+}
+
+pitch_sizes_parts = {
+    'soprano': 30,
+    'alto': 31,
+    'tenor': 30,
+    'bass': 37
+}
 
 
 class ChoralesDataset(Dataset):
-    def __init__(self, root_dir, context_radius=32, transpositions=[0]):
+    def __init__(self, root_dir, context_radius=32):
         self.root_dir = root_dir
         self.context_radius = context_radius
 
         # Make empty intros for each part
         self.data = {
-            'soprano': [torch.zeros((context_radius, part_size))],
-            'tenor': [torch.zeros((context_radius, part_size))],
-            'alto': [torch.zeros((context_radius, part_size))],
-            'bass': [torch.zeros((context_radius, part_size))],
+            'soprano': [torch.zeros((context_radius, pitch_sizes_parts['soprano'] + len(indices_parts)))],
+            'tenor': [torch.zeros((context_radius, pitch_sizes_parts['tenor'] + len(indices_parts)))],
+            'alto': [torch.zeros((context_radius, pitch_sizes_parts['alto'] + len(indices_parts)))],
+            'bass': [torch.zeros((context_radius, pitch_sizes_parts['bass'] + len(indices_parts)))],
             'extra': [torch.zeros((context_radius, len(indices_extra)))]
         }
 
@@ -49,20 +84,7 @@ class ChoralesDataset(Dataset):
         for file_path in sorted(glob(os.path.join(self.root_dir, '*.pt'))):
             data = torch.load(file_path)['data']
             for part_name, part_data in data.items():
-                part_transposed = []
-                for t in transpositions:
-                    if part_name == 'extra':
-                        sharps_offset = (t * 7) % 12
-                        part_cloned = part_data.clone()
-                        part_cloned[:, indices_extra['num_sharps']] = (part_cloned[:, indices_extra['num_sharps']] + sharps_offset) % 12
-                        part_transposed.append(part_cloned)
-                    else:
-                        part_transposed.append(torch.cat([
-                            part_data[:, :len(indices_parts)],
-                            part_data[:, len(indices_parts):].roll(t, dims=1)
-                        ], dim=1))
-                    part_transposed.append(torch.zeros((context_radius, part_data.shape[1])))
-                self.data[part_name].append(torch.cat(part_transposed, dim=0))
+                self.data[part_name].append(torch.cat([part_data, torch.zeros((context_radius, part_data.shape[1]))], dim=0))
 
         for part_name, part_data in self.data.items():
             self.data[part_name] = torch.cat(part_data, dim=0)
@@ -76,8 +98,9 @@ class ChoralesDataset(Dataset):
                    'soprano': self.data['soprano'][idx:idx + 2 * self.context_radius + 1],
                    'alto': self.data['alto'][idx:idx + self.context_radius],
                    'tenor': self.data['tenor'][idx:idx + self.context_radius],
-                   'bass_withcontext': self.data['bass'][idx:idx + 2 * self.context_radius + 1],
                    'bass': self.data['bass'][idx:idx + self.context_radius],
+
+                   'bass_with_context': self.data['bass'][idx:idx + 2 * self.context_radius + 1],
                    'extra': self.data['extra'][idx:idx + 2 * self.context_radius + 1]
                }, {
                    'alto': torch.argmax(self.data['alto'][idx + self.context_radius]),
@@ -92,8 +115,16 @@ def generate_data_inference(time_grid, soprano_path):
     length = math.ceil(stream.highestTime / time_grid)
     data = {
         'extra': torch.zeros((length, len(indices_extra))),
-        'soprano': torch.zeros((length, part_size))
+        'soprano': torch.zeros((length, pitch_sizes_parts['soprano'] + len(indices_parts)))
     }
+
+    # Transpose chorale to C/Am
+    keys = list(stream.flat.getElementsByClass(Key))
+    if len(keys) > 0:
+        transposition = sharps_to_offset[keys[0].sharps]
+        stream.transpose(-transposition, inPlace=True)
+    else:
+        transposition = 0
 
     # Iterate through all musical elements in current voice stream
     for element in stream.flat:
@@ -104,7 +135,7 @@ def generate_data_inference(time_grid, soprano_path):
             if element.duration.quarterLength == 0:
                 continue
 
-            pitch = element.pitch.midi - pitch_size // 2 + len(indices_parts)
+            pitch = element.pitch.midi - offsets_parts['soprano'] + len(indices_parts)
             duration = int(element.duration.quarterLength / time_grid)
 
             # Store pitch and ties
@@ -121,11 +152,18 @@ def generate_data_inference(time_grid, soprano_path):
             data['soprano'][offset + 1:offset + duration, indices_parts['is_continued']] = 1
 
         if type(element) == TimeSignature:
-            data['extra'][offset:, indices_extra['time_numerator']] = element.numerator
-            data['extra'][offset:, indices_extra['time_denominator']] = element.denominator
-
-        if type(element) == KeySignature or type(element) == Key:
-            data['extra'][offset:, indices_extra['num_sharps']] = element.sharps
+            if element.ratioString == '3/4':
+                data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 1
+                data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 0
+                data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 0
+            elif element.ratioString == '4/4':
+                data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 0
+                data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 1
+                data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 0
+            elif element.ratioString == '3/2':
+                data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 0
+                data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 0
+                data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 1
 
     measure_offsets = [o / time_grid for o in stream.measureOffsetMap().keys()]
     cur_offset = stream.flat.notesAndRests[0].beat
@@ -139,7 +177,8 @@ def generate_data_inference(time_grid, soprano_path):
 
     return {
         'data': data,
-        'metadata': stream.metadata
+        'metadata': stream.metadata,
+        'transposition': transposition
     }
 
 
@@ -149,7 +188,7 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
     if os.path.exists(target_dir) and not overwrite:
         return target_dir
 
-    if overwrite:
+    if overwrite and os.path.exists(target_dir):
         shutil.rmtree(root_dir)
 
     train_dir = os.path.join(target_dir, 'train')
@@ -176,6 +215,15 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
         except KeyError:
             continue
 
+        # Save soprano in own file
+        chorale['Soprano'].write('musicxml', os.path.join(musicxml_dir, f'{str(chorale.metadata.number).zfill(3)}_soprano.musicxml'))
+
+        # Transpose chorale to C/Am
+        keys = list(chorale.flat.getElementsByClass(Key))
+        if len(keys) > 0:
+            transposition = sharps_to_offset[keys[0].sharps]
+            chorale.transpose(-transposition, inPlace=True)
+
         length = math.ceil(streams['soprano'].highestTime / time_grid)
         data = {
             'extra': torch.zeros((length, len(indices_extra)))
@@ -183,7 +231,7 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
         for part_name, part in streams.items():
             part = part.flat
             # Init empty tensor for current voice
-            data[part_name] = torch.zeros((length, part_size))
+            data[part_name] = torch.zeros((length, pitch_sizes_parts[part_name] + len(indices_parts)))
 
             # Iterate through all musical elements in current voice stream
             for element in part:
@@ -194,7 +242,7 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
                     if element.duration.quarterLength == 0:
                         continue
 
-                    pitch = element.pitch.midi - pitch_size // 2 + len(indices_parts)
+                    pitch = element.pitch.midi - offsets_parts[part_name] + len(indices_parts)
                     duration = int(element.duration.quarterLength / time_grid)
 
                     # Store pitch and ties
@@ -212,11 +260,18 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
 
                 if part_name == 'soprano':
                     if type(element) == TimeSignature:
-                        data['extra'][offset:, indices_extra['time_numerator']] = element.numerator
-                        data['extra'][offset:, indices_extra['time_denominator']] = element.denominator
-
-                    if type(element) == KeySignature or type(element) == Key:
-                        data['extra'][offset:, indices_extra['num_sharps']] = element.sharps
+                        if element.ratioString == '3/4':
+                            data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 1
+                            data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 0
+                            data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 0
+                        elif element.ratioString == '4/4':
+                            data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 0
+                            data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 1
+                            data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 0
+                        elif element.ratioString == '3/2':
+                            data['extra'][offset:, indices_extra['has_time_signature_3/4']] = 0
+                            data['extra'][offset:, indices_extra['has_time_signature_4/4']] = 0
+                            data['extra'][offset:, indices_extra['has_time_signature_3/2']] = 1
 
         measure_offsets = [o / time_grid for o in streams['soprano'].measureOffsetMap().keys()]
         cur_offset = streams['soprano'].flat.notesAndRests[0].beat
@@ -235,7 +290,6 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
         }, target_file_path)
 
         chorale.write('musicxml', os.path.join(musicxml_dir, f'{str(chorale.metadata.number).zfill(3)}_full.musicxml'))
-        chorale['Soprano'].write('musicxml', os.path.join(musicxml_dir, f'{str(chorale.metadata.number).zfill(3)}_soprano.musicxml'))
 
     # Move files to train / test directories
     file_paths = glob(os.path.join(target_dir, '*.pt'))
@@ -249,12 +303,11 @@ def _generate_data_training(time_grid, root_dir, overwrite, split):
     return target_dir
 
 
-def _make_data_loaders(root_dir, batch_size, num_workers, context_radius, transpositions):
+def _make_data_loaders(root_dir, batch_size, num_workers, context_radius):
     # Training data loader: random sampling
     train_dataset = ChoralesDataset(
         root_dir=os.path.join(root_dir, 'train'),
-        context_radius=context_radius,
-        transpositions=transpositions
+        context_radius=context_radius
     )
     train_sampler = RandomSampler(train_dataset)
     train_batch_sampler = BatchSampler(
@@ -269,8 +322,7 @@ def _make_data_loaders(root_dir, batch_size, num_workers, context_radius, transp
     # Testing data loader: sequential sampling
     test_dataset = ChoralesDataset(
         root_dir=os.path.join(root_dir, 'test'),
-        context_radius=context_radius,
-        transpositions=[0]
+        context_radius=context_radius
     )
     test_sampler = SequentialSampler(test_dataset)
     test_batch_sampler = BatchSampler(
@@ -284,13 +336,11 @@ def _make_data_loaders(root_dir, batch_size, num_workers, context_radius, transp
 
     return {
         'train': train_data_loader,
-        'test': test_data_loader,
-        'input_size': part_size * 4 + len(indices_extra),
-        'output_sizes': part_size
+        'test': test_data_loader
     }
 
 
-def get_data_loaders(time_grid=0.25, root_dir=None, overwrite=False, split=0.15, batch_size=1, num_workers=1, context_radius=32, transpositions=[-6, -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5]):
+def get_data_loaders(time_grid=0.25, root_dir=None, overwrite=False, split=0.15, batch_size=1, num_workers=1, context_radius=32):
     if root_dir is None:
         root_dir = os.path.join('.', 'data')
 
@@ -303,7 +353,6 @@ def get_data_loaders(time_grid=0.25, root_dir=None, overwrite=False, split=0.15,
 
     data_loaders = _make_data_loaders(
         data_dir,
-        transpositions=transpositions,
         batch_size=batch_size,
         num_workers=num_workers,
         context_radius=context_radius
